Refactored AsteroidSystem and added diverse light system
The AsteroidSystem size and properties are shrunk and the unnecessary code was cleared out, making it smaller and streamlining it. For the LightSystem, different types of lights, such as directional, point, and spotlight, were introduced to improve environment representation. This includes adding new structures and adjusting how they are initialized and used within the shader files. Changes were also made to enhance shader performances and accuracy. The update aims to ensure smoother gameplay, stronger light reactions, and more visually appealing graphics.