Add new SimpleDraw methods for models and meshes
A SimpleDraw method was added to models and meshes to allow for simpler drawing commands. This addition also includes the creation of a new depth linearization calculation in the shader script, which will improve the depth precision in the depth map. The changes should help make the rendering process more efficient and improve overall code clarity.