Add UpdateData method to lights and update light data
The UpdateData virtual method has been added to DirLight, SpotLight, and PointLight classes. This improves the organization and maintainability of the code by centralizing the logic for updating the light data. The light data now gets updated, only if it is dirty, during the run-time directly in the LightSystem Update method.