Refactor shadow rendering and light space calculations
The codebase is updated to bring significant changes in lighting, rendering and shadowing subsystems. This includes rearrangement of shadow rendering sequences, enhancements to light space matrix calculations, introduction of dynamic near and far plane definitions, and optimizing the use of shaders. This commit also adds the Capacity to have Point Light in the scene.