Switch point light to spot light; update lighting position and transform
Replaced point light with spot light in the main function to improve light simulation. Adjusted the far plane values in the SpotLight class and updated the UpdateData function in SpotLight.cpp for better 3D light projection. Added 'getGlobalRotation' function to the Transform class, updating the transform parameters according to the light position.