🚀 Added pthreads dependency and warning suppression 🛠️
This commit introduces the pthreads package to the project dependencies in the vcpkg.json file. 👀
It also suppresses warnings in the CMakeLists.txt file to improve the build experience. 👷
In addition, the commit provides enhancements in the LightSystem file by defining new zones for the light sources for better profiling using the Tracy Profiler. 🌟