🔧 Improve Rendering Logic and Add Fog of War Support 🎮
This commit refines the rendering system, specifically by enabling GL_BLEND for the PhongPipeline and modifying the alpha discard threshold in the spriteRender Shader. It also introduces support for 'Fog of War' 🌫 feature in `BetterSpriteRender`, which can be toggled for specific sprites.

In addition, the code commitment removes unnecessary codes and redistributes some responsibilities, like billboard calculations, for better structure and readability. The cursor and camera interactions have also been regulated to only function when the game is started. 🔀📱