🔥 Remove 'pan_gabka_attack.fbx' model file 🔥
This commit involves the removal of the 'pan_gabka_attack.fbx' model file from the res/models/gabka path. The deletion is due to the cleaning and optimization of the project. Now that the file is deleted, resources linked with 'pan_gabka_attack.fbx' might be affected, and adjustments might be needed.